S02E04_multimodal_processing1/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai-whisper>=20231117
Pillow>=10.0.0 